backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.24.1
pydantic==2.4.2
pydantic-settings==2.0.3
openai>=1.0.0
python-jose==3.3.0
ffmpeg-python>=0.2.0
requests==2.31.0
moviepy==1.0.3
opencv-python==4.8.1.78
aiofiles==23.2.1
pytest==7.4.0
pytest-asyncio==0.21.1
# Audio processing
librosa>=0.11.0
# madmom>=0.16.1  # Commented out due to C extension requirements
numpy==1.24.3
cython>=0.25.0
tqdm==4.66.1
huggingface-hub>=0.16.0
transformers>=4.30.0
diffusers>=0.18.0
safetensors>=0.3.1
peft>=0.4.0
xformers>=0.0.20
triton>=2.0.0
replicate==0.15.4 